requirements.txt
beautifulsoup4==4.11.1
python-dotenv==0.20.0
selenium==4.4.2
tinydb==4.7.0
undetected_chromedriver==3.1.5.post4
